pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=1.84.0",
    "pdfplumber>=0.11.6",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
]
